Adding an example driver to be used as a template
The example_driver.hpp has comments to explain the ordering and layout
of the driver as well as a few reasons for why things need to be in the
way they are in the file.

The example driver doesn't actually do anything, and as long as it is
never instatiated, it will not take up space on the exectuable. May need
to move to the docs are later.